Bug 2054: Round Robin code caused H5FFlush to corrupt a file.
John Mainzer fixed the bug and added a test which wrote file and flush a few
time; close the file then open it by serial and read simple structure. I
changed the test to two parallel running parts of ..._writer and ..._reader
and have the reader verify the file after every flush by the writer.

Tested: parallel in Jam and Amani.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@19683 dab4d1a6-ed17-0410-a064-d1ae371a2980