Purpose: Address TOCTOU warnings in h5jam and h5unjam
Description: Coverity is afraid that the state of the input file could change
between the call to stat() and the call to open().  This is called a time-of-
check time-of-use (TOCTOU) vulnerability.  Modified stat calls to fstat which
uses an open file pointer so it (hopefully) won't complain any more.

Tested: Fedora


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@20046 dab4d1a6-ed17-0410-a064-d1ae371a2980