Interim checkin of AIO/VFD code.
Updated AIO/VFD branch from trunk and dealt with bugs and incompatibilities 
as necessary.  Amongst other things, this required deleting a number of files
unrelated to the AIO/VFD work as detailed below.

Modified test/vfd to be express test aware, and to issue notices when skipping
tests.

Added support for the H5_HAVE_AIO #define.  When H5_HAVE_AIO is not defined,
all but the top level AIO code is Eifdef'ed out.

With the help of Mike M, updated the configure code to define H5_HAVE_AIO 
for Linux and freebsd.  H5_HAVE_AIO is undefined in all other cases.

Tested serial and parallel on Phoenix.  Tested serial on jam, heiwa, linew, and
freedom.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@20269 dab4d1a6-ed17-0410-a064-d1ae371a2980