Fixed bug in test/vfd.c (HDFFV-5936) that caused STDIO VFD test to fail on some platforms.
The problem was that the test was a clone of the SEC2 test.  The SEC2 test accepts a file descriptor (int) via a void pointer from H5Fget_vfd_handle() whereas the STDIO test needs to accept a FILE pointer.  Storing the FILE pointer in an integer failed on some systems.  The STDIO test now stores the file handle in FILE pointer and checks it for NULL instead of < 0.

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@20549 dab4d1a6-ed17-0410-a064-d1ae371a2980