Purpose: Improve performance when edge chunk filters are disabled with no chunk cache.
Description:
When edge chunk filters and the chunk cache are disabled, the library would
previously go through the lock/unlock routines then immediately evict the chunk
as is done for filtered datasets, even for the unfiltered edge chunks.  This
forced the entire chunk to be read/written, even if only part of the chunk was
selected.  Modified the chunk cache to detect this case, and read/write only the
selected portion directly from/to the unfiltered chunk.

Tested: jam, amani (h5committest, heiwa down), fedora


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@20555 dab4d1a6-ed17-0410-a064-d1ae371a2980