Purpose: Fix a bug in h5diff when enum values are compared that do          not represent a valid enum value.
Description:
The h5diff code compares enum values by converting them to strings and
then comparing them.  When the enum value is out of range and can't be
converted to an enum string representation, the comparison was skipped.

The code now flags differences when one of the two enum values is out
of range (two out of range values are compared in memory).  A test has
been added to the tools test script.

This fixes JIRA HDFFV-7527

Tested on: jam, koala, heiwa (h5committest)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@20724 dab4d1a6-ed17-0410-a064-d1ae371a2980