Purpose: Fix HDFFV-7522 "--enable-production=xxx will produce incorrect configure summary"
Description:
Fixed a typo in configure.in that resulted in "-enableval" being
displayed by the configure summary (and set in the CONFIG_MODE
makefile variable) if the production mode was set to anything other
than yes, no, or profile. The summary and CONFIG_MODE variable will
now be set to the value specified by the user.

Tested:
jam, h5committest



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@21067 dab4d1a6-ed17-0410-a064-d1ae371a2980