Added interim verision of H5Fget_file_image() and associated test code.
This version has been tested with sec2, stdio, and core file drivers.  It
should work with the family file driver.  It almost certainly will not 
work with the spli/multi file drivers.

We need to decide whether, and if so, how we will support the split/multi
file drivers.  However, this version should be good enough for 
Christian's testing.

Known deficits at this time:

1) must run reconfigure on Jam to add trace calls.

2) must add invalid input rejection test

3) must add test for family file driver

4) must decide how we will address the split/multi file driver 
   issue and implement.

Tested: serial only on phoenix and jam.




git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@21768 dab4d1a6-ed17-0410-a064-d1ae371a2980