Purpose: Fix HDFFV-7833
Description:
When shrinking a chunked dataset, the library fills in the unused parts of
chunks that have been shrunk.  The fill value buffer allocated for this purpose
had a maximum size of 1 MB, but the fill was performed in a single operation.
Therefore, if the amount of unused space in a chunk after being shrunk was
greater than 1 MB, the library would read off the end of the fill value buffer.
Changed the maximum fill buffer size to be equal to the chunk size.

Tested: durandal; jam, koala, heiwa (h5committest)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@21789 dab4d1a6-ed17-0410-a064-d1ae371a2980