Purpose: Add persistent file alignment option
Description:
Added two new functions, H5Pset_persist_alignment and H5Pget_persist_alignment.
These allow the user to set a persistent file alignment, which overrides that
set with H5Pset_alignment, and is is followed every time the file is written to,
even if it has been closed and reopened.  In addition, aligned blocks of data
will "reserve" a multiple of the alignment size, allowing this potentially
largersize to be written to without needing to first read the data in the
"remainder".  This will provide a performance boost to the direct I/O driver.

Files created with this property can still be read by older versions of the
library, but if opened with write acces by an older version will lose the
aforementioned guarantee permanently, and will therefore lose the performance
boost.

Tested: durandal


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@21907 dab4d1a6-ed17-0410-a064-d1ae371a2980