Feature: HDFFV-7793: AIX Fortran and C++ Compiler version information
Added code to display the version information of XL fortran and C++ compiler
version information. (This is just a quick fix for the AIX XL compilers.
The overall algorithm of compilers version information needs an overhaul.)

Tested: ADA machine. (No committest because the changes applied to AIX XL
compilers only.)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@21998 dab4d1a6-ed17-0410-a064-d1ae371a2980