Fix for HDFFV-7769 (HDF5 tests time out on Windows when core VFD is set).
This occurs due to the istore test creating very large files on systems which
do not have POSIX-like sparse file semantics.  The large amount of I/O causes
the test to run for a very long period of time.

The fix was to copy the "big" test's sparse file check and only run the
largest sparse file test when POSIX-like sparse file semantics are found.

Tested on:

jam (nfs)
ostrich (nfs)
loyalty (ufs)
64-bit linux VM (ext4)
64-bit OS-X Lion (hfs, detected as not POSIX sparse)
64-bit Windows 7 (NTFS, detected as not POSIX sparse)

The OS-X failure to pass the sparse check is likely an error and will be
entered as a new bug.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@22103 dab4d1a6-ed17-0410-a064-d1ae371a2980