Bug fix: HDFFV-8105
Description:
The test command in COPY_TESTFILES_TO_TESTDIR used the wrong operator -a which
means and. It is not sufficient to test mere existence as $tstfile could have
been a directory or other non-file things.  If the file is not readable, the
$CP command would fail but its return code was not checked.

Solution:
Get rid of the test command and go ahead to $CP it and then check its returned
code and report errors if detected.

Tested; committested. (note that ostrick still failed on errors committed by
other staff).


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@22594 dab4d1a6-ed17-0410-a064-d1ae371a2980