Bug fix: Mac system has dynamic library name in the form of libxyz.dylib. Changed test_plugin.in to copy library file names in the form of libdynlib{123].*.
Another problem: Even when copy failed in copying the libray files (was in the form of *.so*),
the test still passed.  That was because .libs was included in $HDF5_PLUGIN_PATH and the needed
plug in libraries were generated there.
Fixed by copying the plug in library files to separate directories and seetup HDF5_PLUGIN_PATH
to include them but NOT .libs.

Tested: used desycommittest (duck, emu, jam, koala, ostrich) plus cmakehdf5 in jam.
The cmakehdf5 in jam reported a failure in 
Test #594: H5REPART-h5repart_test ....................................***Failed
Since this failure could not be caused by my change in test_plugin.sh, I reported it to Allen
and proceed to check this in.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@23526 dab4d1a6-ed17-0410-a064-d1ae371a2980