Bring 24612 from trunk:
fix bugs in parallel tests exposed in corner cases when running with 1 or 2 processes.
First bug is in testpar/t_mdset.c, where the test reports an error in addition to skipping the test if there are less than three procs. Fix to just skip the test.
Second bug is in testpar/t_dset.c in actual_io_mode tests, where incorrect expected value for IO mode was set if the number of procs running the test is 1.

tested with h5committest.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@24614 dab4d1a6-ed17-0410-a064-d1ae371a2980