Purpose: Fix bug and improve performance (processing time) for multi-dataset I/O
Description: The previous implementation used very inefficient linear searches
of a skip list in a few places.  In addition, it did not check to see if a chunk
was in the same dataset as the one it was looking for, which would cause a bug.
Added a new skiplist for chunks for each dataset involved in a multi dataset I/O
operation, indexed by chunk index, to fix both of these issues.

Tested: jam


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@24723 dab4d1a6-ed17-0410-a064-d1ae371a2980