Purpose: Fix bug and simplify code for multi-dataset I/O
Description:
When a single element is selected in a chunked dataset,the multi dataset I/O
code wouldn't add the chunk to the skip list.  The prevented it from being
included in collective operations.  The test was only passing because it only
tested single element I/O with point selection, which breaks collective
currently.

Refactored related parts of the code, fixing various potential issues.

Tested: jam


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@24730 dab4d1a6-ed17-0410-a064-d1ae371a2980