Removed the core and core_paged VFDs from the list of SWMR-supported VFDs for now.
Many tests fail with the core VFD, probably due to the lack of a
backing store at key parts of the tests. We'll have to revisit
this idea later. It may need to be shelved until we find time
to rework the tests.

Also fixed VFD issues in the dsets and links test.

'make check-vfd' now passes.

Tested on: jam with 'make check-vfd'


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@25786 dab4d1a6-ed17-0410-a064-d1ae371a2980