Fixes compile failures when building on Windows when UNICODE is defined.
This was due to wrapping a variable with the TEXT() macro and not
forcing the <function>A form of certain Win32 API calls are always used.

Although we do not defined UNICODE in our builds, others have encountered
problems when building HDF5 as a part of a larger project where UNICODE
is defined.

Fix for: HDFFV-8927

Tested on: 64-bit Windows 7, Visual Studio 2012, CMake 3.2
           All combinations of static/shared and UNICODE defined/undefined
           NOTE: plugin tests are skipped with static builds

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@26273 dab4d1a6-ed17-0410-a064-d1ae371a2980