Partial revert of r26347.
Since off_t is exposed by the C API, the large file definitions MUST
be exposed so that client code matches the library.

These definitions have been switched back to AM_CPPFLAGS, which is
exported to h5cc, etc. The POSIX and GNU definitions changes from
r26347 have been left as using H5_CPPFLAGS, which is not sent to
h5cc as we don't want to inflict our needs on external users.

Fixes: HDFFV-9152

Tested on: h5committest



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@26363 dab4d1a6-ed17-0410-a064-d1ae371a2980