Fixes to production mode failures.
The basic problem was an attempt to remove light sanity 
checks in the metadata cache that run even in production
mode.  This ran into problems as the sanity checking code
in the metadata cache was designed to run independently
of the NDEBUG #define.

While it is possible to convert this code for use only 
when NDEBUG is defined, this would take a bit of doing,
and I doubt that it would result in any detectable 
speedup.

Thus, for the time being at least, I restored the status
quo ante.

Tested on phoenix (serial, serial/production, & parallel)
          jam (serial)



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@26404 dab4d1a6-ed17-0410-a064-d1ae371a2980