Added an empty VDS testfile generator to tools/misc directory.
test/Makefile.in was updated due to changes in the Makefile.am
file that missed a reconfigure. The changes just added a few
VDS test files to the deletion list.

Note that there are trace issues with a new VDS struct that
show up when running bin/reconfigure. I reverted these for
fixing at a future date.

Tested on: jam


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@27150 dab4d1a6-ed17-0410-a064-d1ae371a2980