SWMR-26: make check-vfd runs for a long time The problem is in test/dsets which infinite looped when HDF5_DRIVER is "sec2" because it had a for loop using a bool type variable (compress) but the loop tried to increment to a third value. Changed the type of compress to int allows the loop to run properly.
Secondly, the resas why it did not infinite loop when not in check-vfd is because the
getenv(HDF5_DRIVER) mistakenly assign driver as "nomatch" if it is not defined.
dsets skipped the test because "nomatch" is not one of the supported driver.
Fixed the code to assign "sec2" if $HDF5_DRIVER is not defined.

Tested: kituo, ostrich, osx1010dev


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@28321 dab4d1a6-ed17-0410-a064-d1ae371a2980